[TG-TECH][refactor]  refactored header
