+++ src/app/core/header/header.component.html
@@ -2,7 +2,7 @@
     <nav class="main-nav">
         <div class="container">
             <a class="logo"
-               [routerLink]="[mainPageRoute.name]"
+               [routerLink]="[mainPageRoute.path]"
                [style.background-image]="'url(' + mainPageRoute.iconImage + ')'">Smart House</a>
             <ul id="nav-mobile" class="right hide-on-med-and-down">
                 <li *ngFor="let route of navigationRoutes">
